refactor: extract scan and factorize functions to separate modules
- Create flox/scan.py with groupby_scan and related functions
- Create flox/factorize.py with factorize_ and related functions
- Move scan-related types to types.py under TYPE_CHECKING
- Update imports across codebase (core, tests, benchmarks)
- Maintain backward compatibility through __init__.py exports

This improves code organization by separating scan operations and
factorization logic from the large core.py module while preserving
all existing functionality.
