+++ flox/core.py
@@ -1,18 +1,15 @@
 from __future__ import annotations
 
-import copy
 import datetime
 import itertools
 import logging
 import math
 import operator
 import warnings
-from collections import namedtuple
 from collections.abc import Callable, Sequence
 from concurrent.futures import ThreadPoolExecutor
 from dataclasses import dataclass
-from enum import Enum, auto
-from functools import partial, reduce
+from functools import partial
 from itertools import product
 from numbers import Integral
 from typing import (
@@ -20,10 +17,8 @@ from typing import (
     Any,
     Literal,
     TypeAlias,
-    TypedDict,
     TypeVar,
     Union,
-    cast,
     overload,
 )
 
@@ -36,18 +31,17 @@ from scipy.sparse import csc_array, csr_array
 from . import xrdtypes
 from .aggregate_flox import _prepare_for_flox
 from .aggregations import (
-    AGGREGATIONS,
     Aggregation,
-    AlignedArrays,
-    Scan,
-    ScanState,
     _atleast_1d,
     _initialize_aggregation,
     generic_aggregate,
     quantile_new_dims_func,
 )
 from .cache import memoize
+from .factorize import _factorize_multiple, factorize_
 from .lib import ArrayLayer, dask_array_type, sparse_array_type
+from .reindex import reindex_
+from .utils import ReindexArrayType
 from .xrutils import (
     _contains_cftime_datetimes,
     _to_pytimedelta,
@@ -88,7 +82,6 @@ if TYPE_CHECKING:
     T_Func = str | Callable
     T_Funcs = T_Func | Sequence[T_Func]
     T_Agg = str | Aggregation
-    T_Scan = str | Scan
     T_Axis = int
     T_Axes = tuple[T_Axis, ...]
     T_AxesOpt = T_Axis | T_Axes | None
@@ -104,7 +97,6 @@ T = TypeVar("T")
 
 IntermediateDict = dict[str | Callable, Any]
 FinalResultsDict = dict[str, Union["DaskArray", "CubedArray", np.ndarray]]
-FactorProps = namedtuple("FactorProps", "offset_group nan_sentinel nanmask")
 
 # This dummy axis is inserted using np.expand_dims
 # and then reduced over during the combine stage by
@@ -112,36 +104,6 @@ FactorProps = namedtuple("FactorProps", "offset_group nan_sentinel nanmask")
 DUMMY_AXIS = -2
 
 logger = logging.getLogger("flox")
-
-
-class ReindexArrayType(Enum):
-    """
-    Enum describing which array type to reindex to.
-
-    These are enumerated, rather than accepting a constructor,
-    because we might want to optimize for specific array types,
-    and because they don't necessarily have the same signature.
-
-    For example, scipy.sparse.COO only supports a fill_value of 0.
-    """
-
-    AUTO = auto()
-    NUMPY = auto()
-    SPARSE_COO = auto()
-    # Sadly, scipy.sparse.coo_array only supports fill_value = 0
-    # SCIPY_SPARSE_COO = auto()
-    # SPARSE_GCXS = auto()
-
-    def is_same_type(self, other) -> bool:
-        match self:
-            case ReindexArrayType.AUTO:
-                return True
-            case ReindexArrayType.NUMPY:
-                return isinstance(other, np.ndarray)
-            case ReindexArrayType.SPARSE_COO:
-                import sparse
-
-                return isinstance(other, sparse.COO)
 
 
 @dataclass
@@ -181,16 +143,6 @@ class ReindexStrategy:
             import sparse
 
             return sparse.COO.from_numpy(np.ones(shape=(0,) * other.ndim, dtype=dtype), fill_value=fill_value)
-
-
-class FactorizeKwargs(TypedDict, total=False):
-    """Used in _factorize_multiple"""
-
-    by: T_Bys
-    axes: T_Axes
-    fastpath: bool
-    reindex: bool
-    sort: bool
 
 
 def _postprocess_numbagg(result, *, func, fill_value, size, seen_groups):
@@ -741,316 +693,6 @@ def rechunk_for_blockwise(array: DaskArray, axis: T_Axis, labels: np.ndarray) ->
         return array.rechunk({axis: newchunks})
 
 
-def reindex_numpy(array, from_: pd.Index, to: pd.Index, fill_value, dtype, axis: int):
-    idx = from_.get_indexer(to)
-    indexer = [slice(None, None)] * array.ndim
-    indexer[axis] = idx
-    reindexed = array[tuple(indexer)]
-    if (idx == -1).any():
-        if fill_value is None:
-            raise ValueError("Filling is required. fill_value cannot be None.")
-        indexer[axis] = idx == -1
-        reindexed = reindexed.astype(dtype, copy=False)
-        reindexed[tuple(indexer)] = fill_value
-    return reindexed
-
-
-def reindex_pydata_sparse_coo(array, from_: pd.Index, to: pd.Index, fill_value, dtype, axis: int):
-    import sparse
-
-    assert axis == -1
-
-    # Are there any elements in `to` that are not in `from_`.
-    if isinstance(to, pd.RangeIndex) and len(to) > len(from_):
-        # 1. pandas optimizes set difference between two RangeIndexes only
-        # 2. We want to avoid realizing a very large numpy array in to memory.
-        #    This happens in the `else` clause.
-        #    There are potentially other tricks we can play, but this is a simple
-        #    and effective one. If a user is reindexing to sparse, then len(to) is
-        #    almost guaranteed to be > len(from_). If len(to) <= len(from_), then realizing
-        #    another array of the same shape should be fine.
-        needs_reindex = True
-    else:
-        needs_reindex = (from_.get_indexer(to) == -1).any()
-
-    if needs_reindex and fill_value is None:
-        raise ValueError("Filling is required. fill_value cannot be None.")
-
-    idx = to.get_indexer(from_)
-    mask = idx != -1  # indices along last axis to keep
-    if mask.all():
-        mask = slice(None)
-    shape = array.shape
-
-    if isinstance(array, sparse.COO):
-        subset = array[..., mask]
-        data = subset.data
-        coords = subset.coords
-        if subset.nnz > 0:
-            coords[-1, :] = idx[mask][coords[-1, :]]
-        if fill_value is None:
-            # no reindexing is actually needed (dense case)
-            # preserve the fill_value
-            fill_value = array.fill_value
-    else:
-        ranges = np.broadcast_arrays(
-            *np.ix_(*(tuple(np.arange(size) for size in shape[:axis]) + (idx[mask],)))
-        )
-        coords = np.stack(ranges, axis=0).reshape(array.ndim, -1)
-        data = array[..., mask].reshape(-1)
-
-    reindexed = sparse.COO(
-        coords=coords,
-        data=data.astype(dtype, copy=False),
-        shape=(*array.shape[:axis], to.size),
-        fill_value=fill_value,
-    )
-
-    return reindexed
-
-
-def reindex_(
-    array: np.ndarray,
-    from_,
-    to,
-    *,
-    array_type: ReindexArrayType = ReindexArrayType.AUTO,
-    fill_value: Any = None,
-    axis: T_Axis = -1,
-    promote: bool = False,
-) -> np.ndarray:
-    if not isinstance(to, pd.Index):
-        if promote:
-            to = pd.Index(to)
-        else:
-            raise ValueError("reindex requires a pandas.Index or promote=True")
-
-    if to.ndim > 1:
-        raise ValueError(f"Cannot reindex to a multidimensional array: {to}")
-
-    if array.shape[axis] == 0:
-        # all groups were NaN
-        shape = array.shape[:-1] + (len(to),)
-        if array_type in (ReindexArrayType.AUTO, ReindexArrayType.NUMPY):
-            reindexed = np.full(shape, fill_value, dtype=array.dtype)
-        else:
-            raise NotImplementedError
-        return reindexed
-
-    from_ = pd.Index(from_)
-    # short-circuit for trivial case
-    if from_.equals(to) and array_type.is_same_type(array):
-        return array
-
-    if from_.dtype.kind == "O" and isinstance(from_[0], tuple):
-        raise NotImplementedError(
-            "Currently does not support reindexing with object arrays of tuples. "
-            "These occur when grouping by multi-indexed variables in xarray."
-        )
-    if fill_value is xrdtypes.NA or isnull(fill_value):
-        new_dtype, fill_value = xrdtypes.maybe_promote(array.dtype)
-    else:
-        new_dtype = array.dtype
-
-    if array_type is ReindexArrayType.AUTO:
-        if isinstance(array, sparse_array_type):
-            array_type = ReindexArrayType.SPARSE_COO
-        else:
-            # TODO: generalize here
-            # Right now, we effectively assume NEP-18 I think
-            array_type = ReindexArrayType.NUMPY
-
-    if array_type is ReindexArrayType.NUMPY:
-        reindexed = reindex_numpy(array, from_, to, fill_value, new_dtype, axis)
-    elif array_type is ReindexArrayType.SPARSE_COO:
-        reindexed = reindex_pydata_sparse_coo(array, from_, to, fill_value, new_dtype, axis)
-    return reindexed
-
-
-def offset_labels(labels: np.ndarray, ngroups: int) -> tuple[np.ndarray, int]:
-    """
-    Offset group labels by dimension. This is used when we
-    reduce over a subset of the dimensions of by. It assumes that the reductions
-    dimensions have been flattened in the last dimension
-    Copied from xhistogram &
-    https://stackoverflow.com/questions/46256279/bin-elements-per-row-vectorized-2d-bincount-for-numpy
-    """
-    assert labels.ndim > 1
-    offset: np.ndarray = (
-        labels + np.arange(math.prod(labels.shape[:-1])).reshape((*labels.shape[:-1], -1)) * ngroups
-    )
-    # -1 indicates NaNs. preserve these otherwise we aggregate in the wrong groups!
-    offset[labels == -1] = -1
-    size: int = math.prod(labels.shape[:-1]) * ngroups
-    return offset, size
-
-
-def _factorize_single(by, expect, *, sort: bool, reindex: bool) -> tuple[pd.Index, np.ndarray]:
-    flat = by.reshape(-1)
-    if isinstance(expect, pd.RangeIndex):
-        # idx is a view of the original `by` array
-        # copy here so we don't have a race condition with the
-        # group_idx[nanmask] = nan_sentinel assignment later
-        # this is important in shared-memory parallelism with dask
-        # TODO: figure out how to avoid this
-        idx = flat.copy()
-        found_groups = cast(pd.Index, expect)
-        # TODO: fix by using masked integers
-        idx[idx > expect[-1]] = -1
-
-    elif isinstance(expect, pd.IntervalIndex):
-        if expect.closed == "both":
-            raise NotImplementedError
-        bins = np.concatenate([expect.left.to_numpy(), expect.right.to_numpy()[[-1]]])
-
-        # digitize is 0 or idx.max() for values outside the bounds of all intervals
-        # make it behave like pd.cut which uses -1:
-        if len(bins) > 1:
-            right = expect.closed_right
-            idx = np.digitize(
-                flat,
-                bins=bins.view(np.int64) if bins.dtype.kind == "M" else bins,
-                right=right,
-            )
-            idx -= 1
-            within_bins = flat <= bins.max() if right else flat < bins.max()
-            idx[~within_bins] = -1
-        else:
-            idx = np.zeros_like(flat, dtype=np.intp) - 1
-        found_groups = cast(pd.Index, expect)
-    else:
-        if expect is not None and reindex:
-            sorter = np.argsort(expect)
-            groups = expect[(sorter,)] if sort else expect
-            idx = np.searchsorted(expect, flat, sorter=sorter)
-            mask = ~np.isin(flat, expect) | isnull(flat) | (idx == len(expect))
-            if not sort:
-                # idx is the index in to the sorted array.
-                # if we didn't want sorting, unsort it back
-                idx[(idx == len(expect),)] = -1
-                idx = sorter[(idx,)]
-            idx[mask] = -1
-        else:
-            idx, groups = pd.factorize(flat, sort=sort)
-        found_groups = cast(pd.Index, groups)
-
-    return (found_groups, idx.reshape(by.shape))
-
-
-def _ravel_factorized(*factorized: np.ndarray, grp_shape: tuple[int, ...]) -> np.ndarray:
-    group_idx = np.ravel_multi_index(factorized, grp_shape, mode="wrap")
-    # NaNs; as well as values outside the bins are coded by -1
-    # Restore these after the raveling
-    nan_by_mask = reduce(np.logical_or, [(f == -1) for f in factorized])
-    group_idx[nan_by_mask] = -1
-    return group_idx
-
-
-@overload
-def factorize_(
-    by: T_Bys,
-    axes: T_Axes,
-    *,
-    fastpath: Literal[True],
-    expected_groups: T_ExpectIndexOptTuple | None = None,
-    reindex: bool = False,
-    sort: bool = True,
-) -> tuple[np.ndarray, tuple[pd.Index, ...], tuple[int, ...], int, int, None]: ...
-
-
-@overload
-def factorize_(
-    by: T_Bys,
-    axes: T_Axes,
-    *,
-    expected_groups: T_ExpectIndexOptTuple | None = None,
-    reindex: bool = False,
-    sort: bool = True,
-    fastpath: Literal[False] = False,
-) -> tuple[np.ndarray, tuple[pd.Index, ...], tuple[int, ...], int, int, FactorProps]: ...
-
-
-@overload
-def factorize_(
-    by: T_Bys,
-    axes: T_Axes,
-    *,
-    expected_groups: T_ExpectIndexOptTuple | None = None,
-    reindex: bool = False,
-    sort: bool = True,
-    fastpath: bool = False,
-) -> tuple[np.ndarray, tuple[pd.Index, ...], tuple[int, ...], int, int, FactorProps | None]: ...
-
-
-def factorize_(
-    by: T_Bys,
-    axes: T_Axes,
-    *,
-    expected_groups: T_ExpectIndexOptTuple | None = None,
-    reindex: bool = False,
-    sort: bool = True,
-    fastpath: bool = False,
-) -> tuple[np.ndarray, tuple[pd.Index, ...], tuple[int, ...], int, int, FactorProps | None]:
-    """
-    Returns an array of integer codes for groups (and associated data)
-    by wrapping pd.cut and pd.factorize (depending on isbin).
-    This method handles reindex and sort so that we don't spend time reindexing / sorting
-    a possibly large results array. Instead we set up the appropriate integer codes (group_idx)
-    so that the results come out in the appropriate order.
-    """
-    if expected_groups is None:
-        expected_groups = (None,) * len(by)
-
-    if len(by) > 2:
-        with ThreadPoolExecutor() as executor:
-            futures = [
-                executor.submit(partial(_factorize_single, sort=sort, reindex=reindex), groupvar, expect)
-                for groupvar, expect in zip(by, expected_groups)
-            ]
-            results = tuple(f.result() for f in futures)
-    else:
-        results = tuple(
-            _factorize_single(groupvar, expect, sort=sort, reindex=reindex)
-            for groupvar, expect in zip(by, expected_groups)
-        )
-    found_groups = tuple(r[0] for r in results)
-    factorized = [r[1] for r in results]
-
-    grp_shape = tuple(len(grp) for grp in found_groups)
-    ngroups = math.prod(grp_shape)
-    if len(by) > 1:
-        group_idx = _ravel_factorized(*factorized, grp_shape=grp_shape)
-    else:
-        (group_idx,) = factorized
-
-    if fastpath:
-        return group_idx, found_groups, grp_shape, ngroups, ngroups, None
-
-    if len(axes) == 1 and by[0].ndim > 1:
-        # Not reducing along all dimensions of by
-        # this is OK because for 3D by and axis=(1,2),
-        # we collapse to a 2D by and axis=-1
-        offset_group = True
-        group_idx, size = offset_labels(group_idx.reshape(by[0].shape), ngroups)
-    else:
-        size = ngroups
-        offset_group = False
-
-    # numpy_groupies cannot deal with group_idx = -1
-    # so we'll add use ngroups as the sentinel
-    # note we cannot simply remove the NaN locations;
-    # that would mess up argmax, argmin
-    nan_sentinel = size if offset_group else ngroups
-    nanmask = group_idx == -1
-    if nanmask.any():
-        # bump it up so there's a place to assign values to the nan_sentinel index
-        size += 1
-    group_idx[nanmask] = nan_sentinel
-
-    props = FactorProps(offset_group, nan_sentinel, nanmask)
-    return group_idx, tuple(found_groups), grp_shape, ngroups, size, props
-
-
 def chunk_argreduce(
     array_plus_idx: tuple[np.ndarray, ...],
     by: np.ndarray,
@@ -1468,24 +1110,6 @@ def _conc2(x_chunk, key1, key2=slice(None), axis: T_Axes | None = None) -> np.nd
     # for _ in range(mapped[0].ndim-1):
     #    mapped = [mapped]
     # return concatenate3(mapped)
-
-
-def reindex_intermediates(
-    x: IntermediateDict, agg: Aggregation, unique_groups, array_type
-) -> IntermediateDict:
-    new_shape = x["groups"].shape[:-1] + (len(unique_groups),)
-    newx: IntermediateDict = {"groups": np.broadcast_to(unique_groups, new_shape)}
-    newx["intermediates"] = tuple(
-        reindex_(
-            v,
-            from_=np.atleast_1d(x["groups"].squeeze()),
-            to=pd.Index(unique_groups),
-            fill_value=f,
-            array_type=array_type,
-        )
-        for v, f in zip(x["intermediates"], agg.fill_value["intermediate"])
-    )
-    return newx
 
 
 def listify_groups(x: IntermediateDict):
@@ -2315,68 +1939,6 @@ def _convert_expected_groups_to_index(
     return tuple(out)
 
 
-def _lazy_factorize_wrapper(*by: T_By, **kwargs) -> np.ndarray:
-    group_idx, *_ = factorize_(by, **kwargs)
-    return group_idx
-
-
-def _factorize_multiple(
-    by: T_Bys,
-    expected_groups: T_ExpectIndexOptTuple,
-    any_by_dask: bool,
-    sort: bool = True,
-) -> tuple[tuple[np.ndarray], tuple[pd.Index, ...], tuple[int, ...]]:
-    kwargs: FactorizeKwargs = dict(
-        axes=(),  # always (), we offset later if necessary.
-        fastpath=True,
-        # This is the only way it makes sense I think.
-        # reindex controls what's actually allocated in chunk_reduce
-        # At this point, we care about an accurate conversion to codes.
-        reindex=True,
-        sort=sort,
-    )
-    if any_by_dask:
-        import dask.array
-
-        from . import dask_array_ops  # noqa
-
-        # unifying chunks will make sure all arrays in `by` are dask arrays
-        # with compatible chunks, even if there was originally a numpy array
-        inds = tuple(range(by[0].ndim))
-        for by_, expect in zip(by, expected_groups):
-            if expect is None and is_duck_dask_array(by_):
-                raise ValueError("Please provide expected_groups when grouping by a dask array.")
-
-        found_groups = tuple(
-            pd.Index(pd.unique(by_.reshape(-1))) if expect is None else expect
-            for by_, expect in zip(by, expected_groups)
-        )
-        grp_shape = tuple(map(len, found_groups))
-
-        chunks, by_chunked = dask.array.unify_chunks(*itertools.chain(*zip(by, (inds,) * len(by))))
-        group_idxs = [
-            dask.array.map_blocks(
-                _lazy_factorize_wrapper,
-                by_,
-                expected_groups=(expect_,),
-                meta=np.array((), dtype=np.int64),
-                **kwargs,
-            )
-            for by_, expect_ in zip(by_chunked, expected_groups)
-        ]
-        # This could be avoied but we'd use `np.where`
-        # instead `_ravel_factorized` instead i.e. a copy.
-        group_idx = dask.array.map_blocks(
-            _ravel_factorized, *group_idxs, grp_shape=grp_shape, chunks=tuple(chunks.values()), dtype=np.int64
-        )
-
-    else:
-        kwargs["by"] = by
-        group_idx, found_groups, grp_shape, *_ = factorize_(**kwargs, expected_groups=expected_groups)
-
-    return (group_idx,), found_groups, grp_shape
-
-
 @overload
 def _validate_expected_groups(nby: int, expected_groups: None) -> tuple[None, ...]: ...
 
@@ -2953,281 +2515,18 @@ def groupby_reduce(
     return (result, *groups)
 
 
-def groupby_scan(
-    array: np.ndarray | DaskArray,
-    *by: T_By,
-    func: T_Scan,
-    expected_groups: T_ExpectedGroupsOpt = None,
-    axis: int | tuple[int] = -1,
-    dtype: np.typing.DTypeLike = None,
-    method: T_MethodOpt = None,
-    engine: T_EngineOpt = None,
-) -> np.ndarray | DaskArray:
-    """
-    GroupBy reductions using parallel scans for dask.array
-
-    Parameters
-    ----------
-    array : ndarray or DaskArray
-        Array to be reduced, possibly nD
-    *by : ndarray or DaskArray
-        Array of labels to group over. Must be aligned with ``array`` so that
-        ``array.shape[-by.ndim :] == by.shape`` or any disagreements in that
-        equality check are for dimensions of size 1 in `by`.
-    func : {"nancumsum", "ffill", "bfill"} or Scan
-        Single function name or a Scan instance
-    expected_groups : (optional) Sequence
-        Expected unique labels.
-    axis : None or int or Sequence[int], optional
-        If None, reduce across all dimensions of by
-        Else, reduce across corresponding axes of array
-        Negative integers are normalized using array.ndim.
-    fill_value : Any
-        Value to assign when a label in ``expected_groups`` is not present.
-    dtype : data-type , optional
-        DType for the output. Can be anything that is accepted by ``np.dtype``.
-    method : {"blockwise", "cohorts"}, optional
-        Strategy for reduction of dask arrays only:
-          * ``"blockwise"``:
-            Only scan using blockwise and avoid aggregating blocks
-            together. Useful for resampling-style groupby problems where group
-            members are always together. If  `by` is 1D,  `array` is automatically
-            rechunked so that chunk boundaries line up with group boundaries
-            i.e. each block contains all members of any group present
-            in that block. For nD `by`, you must make sure that all members of a group
-            are present in a single block.
-          * ``"cohorts"``:
-            Finds group labels that tend to occur together ("cohorts"),
-            indexes out cohorts and reduces that subset using "map-reduce",
-            repeat for all cohorts. This works well for many time groupings
-            where the group labels repeat at regular intervals like 'hour',
-            'month', dayofyear' etc. Optimize chunking ``array`` for this
-            method by first rechunking using ``rechunk_for_cohorts``
-            (for 1D ``by`` only).
-    engine : {"flox", "numpy", "numba", "numbagg"}, optional
-        Algorithm to compute the groupby reduction on non-dask arrays and on each dask chunk:
-          * ``"numpy"``:
-            Use the vectorized implementations in ``numpy_groupies.aggregate_numpy``.
-            This is the default choice because it works for most array types.
-          * ``"flox"``:
-            Use an internal implementation where the data is sorted so that
-            all members of a group occur sequentially, and then numpy.ufunc.reduceat
-            is to used for the reduction. This will fall back to ``numpy_groupies.aggregate_numpy``
-            for a reduction that is not yet implemented.
-          * ``"numba"``:
-            Use the implementations in ``numpy_groupies.aggregate_numba``.
-          * ``"numbagg"``:
-            Use the reductions supported by ``numbagg.grouped``. This will fall back to ``numpy_groupies.aggregate_numpy``
-            for a reduction that is not yet implemented.
-
-    Returns
-    -------
-    result
-        Aggregated result
-
-    See Also
-    --------
-    xarray.xarray_reduce
-    """
-
-    axis = _atleast_1d(axis)
-    if len(axis) > 1:
-        raise NotImplementedError("Scans are only supported along a single dimension.")
-
-    bys: T_Bys = tuple(np.asarray(b) if not is_duck_array(b) else b for b in by)
-    nby = len(by)
-    by_is_dask = tuple(is_duck_dask_array(b) for b in bys)
-    any_by_dask = any(by_is_dask)
-
-    axis_ = normalize_axis_tuple(axis, array.ndim)
-
-    if engine is not None:
-        raise NotImplementedError("Setting `engine` is not supported for scans yet.")
-    if method is not None:
-        raise NotImplementedError("Setting `method` is not supported for scans yet.")
-    if engine is None:
-        engine = "flox"
-    assert engine == "flox"
-
-    if not is_duck_array(array):
-        array = np.asarray(array)
-
-    if isinstance(func, str):
-        agg = AGGREGATIONS[func]
-    assert isinstance(agg, Scan)
-    agg = copy.deepcopy(agg)
-
-    if (agg == AGGREGATIONS["ffill"] or agg == AGGREGATIONS["bfill"]) and array.dtype.kind != "f":
-        # nothing to do, no NaNs!
-        return array
-
-    if expected_groups is not None:
-        raise NotImplementedError("Setting `expected_groups` and binning is not supported yet.")
-    expected_groups = _validate_expected_groups(nby, expected_groups)
-    expected_groups = _convert_expected_groups_to_index(expected_groups, isbin=(False,) * nby, sort=False)
-
-    # Don't factorize early only when
-    # grouping by dask arrays, and not having expected_groups
-    factorize_early = not (
-        # can't do it if we are grouping by dask array but don't have expected_groups
-        any(is_dask and ex_ is None for is_dask, ex_ in zip(by_is_dask, expected_groups))
-    )
-    if factorize_early:
-        bys, final_groups, grp_shape = _factorize_multiple(
-            bys,
-            expected_groups,
-            any_by_dask=any_by_dask,
-            sort=False,
+def reindex_intermediates(x, agg, unique_groups, array_type):
+    """Reindex intermediate results for groupby operations."""
+    new_shape = x["groups"].shape[:-1] + (len(unique_groups),)
+    newx = {"groups": np.broadcast_to(unique_groups, new_shape)}
+    newx["intermediates"] = tuple(
+        reindex_(
+            v,
+            from_=np.atleast_1d(x["groups"].squeeze()),
+            to=pd.Index(unique_groups),
+            fill_value=f,
+            array_type=array_type,
         )
-    else:
-        raise NotImplementedError
-
-    assert len(bys) == 1
-    by_: np.ndarray
-    (by_,) = bys
-    has_dask = is_duck_dask_array(array) or is_duck_dask_array(by_)
-
-    if array.dtype.kind in "Mm":
-        cast_to = array.dtype
-        array = array.view(np.int64)
-    elif array.dtype.kind == "b":
-        array = array.view(np.int8)
-        cast_to = None
-        if agg.preserves_dtype:
-            cast_to = bool
-    else:
-        cast_to = None
-
-    # TODO: move to aggregate_npg.py
-    if agg.name in ["cumsum", "nancumsum"] and array.dtype.kind in ["i", "u"]:
-        # https://numpy.org/doc/stable/reference/generated/numpy.cumsum.html
-        # it defaults to the dtype of a, unless a
-        # has an integer dtype with a precision less than that of the default platform integer.
-        if array.dtype.kind == "i":
-            agg.dtype = np.result_type(array.dtype, np.int_)
-        elif array.dtype.kind == "u":
-            agg.dtype = np.result_type(array.dtype, np.uint)
-    else:
-        agg.dtype = array.dtype if dtype is None else dtype
-    agg.identity = xrdtypes._get_fill_value(agg.dtype, agg.identity)
-
-    (single_axis,) = axis_  # type: ignore[misc]
-    # avoid some roundoff error when we can.
-    if by_.shape[-1] == 1 or by_.shape == grp_shape:
-        array = array.astype(agg.dtype)
-        if cast_to is not None:
-            array = array.astype(cast_to)
-        return array
-
-    # Made a design choice here to have `preprocess` handle both array and group_idx
-    # Example: for reversing, we need to reverse the whole array, not just reverse
-    #          each block independently
-    inp = AlignedArrays(array=array, group_idx=by_)
-    if agg.preprocess:
-        inp = agg.preprocess(inp)
-
-    if not has_dask:
-        final_state = chunk_scan(inp, axis=single_axis, agg=agg, dtype=agg.dtype)
-        result = _finalize_scan(final_state, dtype=agg.dtype)
-    else:
-        result = dask_groupby_scan(inp.array, inp.group_idx, axes=axis_, agg=agg)
-
-    # Made a design choice here to have `postprocess` handle both array and group_idx
-    out = AlignedArrays(array=result, group_idx=by_)
-    if agg.finalize:
-        out = agg.finalize(out)
-
-    if cast_to is not None:
-        return out.array.astype(cast_to)
-    return out.array
-
-
-def chunk_scan(inp: AlignedArrays, *, axis: int, agg: Scan, dtype=None, keepdims=None) -> ScanState:
-    assert axis == inp.array.ndim - 1
-
-    # I don't think we need to re-factorize here unless we are grouping by a dask array
-    accumulated = generic_aggregate(
-        inp.group_idx,
-        inp.array,
-        axis=axis,
-        engine="flox",
-        func=agg.scan,
-        dtype=dtype,
-        fill_value=agg.identity,
+        for v, f in zip(x["intermediates"], agg.fill_value["intermediate"])
     )
-    result = AlignedArrays(array=accumulated, group_idx=inp.group_idx)
-    return ScanState(result=result, state=None)
-
-
-def grouped_reduce(inp: AlignedArrays, *, agg: Scan, axis: int, keepdims=None) -> ScanState:
-    assert axis == inp.array.ndim - 1
-    reduced = chunk_reduce(
-        inp.array,
-        inp.group_idx,
-        func=(agg.reduction,),
-        axis=axis,
-        engine="flox",
-        dtype=inp.array.dtype,
-        fill_value=agg.identity,
-        expected_groups=None,
-    )
-    return ScanState(
-        state=AlignedArrays(array=reduced["intermediates"][0], group_idx=reduced["groups"]),
-        result=None,
-    )
-
-
-def _zip(group_idx: np.ndarray, array: np.ndarray) -> AlignedArrays:
-    return AlignedArrays(group_idx=group_idx, array=array)
-
-
-def _finalize_scan(block: ScanState, dtype) -> np.ndarray:
-    assert block.result is not None
-    return block.result.array.astype(dtype, copy=False)
-
-
-def dask_groupby_scan(array, by, axes: T_Axes, agg: Scan) -> DaskArray:
-    from dask.array import map_blocks
-    from dask.array.reductions import cumreduction as scan
-
-    from flox.aggregations import scan_binary_op
-
-    if len(axes) > 1:
-        raise NotImplementedError("Scans are only supported along a single axis.")
-    (axis,) = axes
-
-    array, by = _unify_chunks(array, by)
-
-    # 1. zip together group indices & array
-    zipped = map_blocks(
-        _zip,
-        by,
-        array,
-        dtype=array.dtype,
-        meta=array._meta,
-        name="groupby-scan-preprocess",
-    )
-
-    scan_ = partial(chunk_scan, agg=agg)
-    # dask tokenizing error workaround
-    scan_.__name__ = scan_.func.__name__  # type: ignore[attr-defined]
-
-    # 2. Run the scan
-    accumulated = scan(
-        func=scan_,
-        binop=partial(scan_binary_op, agg=agg),
-        ident=agg.identity,
-        x=zipped,
-        axis=axis,
-        # TODO: support method="sequential" here.
-        method="blelloch",
-        preop=partial(grouped_reduce, agg=agg),
-        dtype=agg.dtype,
-    )
-
-    # 3. Unzip and extract the final result array, discard groups
-    result = map_blocks(partial(_finalize_scan, dtype=agg.dtype), accumulated, dtype=agg.dtype)
-
-    assert result.chunks == array.chunks
-
-    return result
+    return newx
